@@ -1,3 +1,19 @@
+const server = require('./src/app.js');
+const { conn } = require('./src/db.js');
+
+// Syncing all the models at once.
+conn.sync({ alter: true }).then(() => {
+  server.listen(3001, () => {
+    console.log('%s listening at 3001'); // eslint-disable-line no-console
+  });
+});
+
+
+
+
+
+
+
 //                       _oo0oo_
 //                      o8888888o
 //                      88" . "88
@@ -17,12 +33,3 @@
 //     =====`-.____`.___ \_____/___.-`___.-'=====
 //                       `=---='
 //     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-const server = require('./src/app.js');
-const { conn } = require('./src/db.js');
-
-// Syncing all the models at once.
-conn.sync({ force: true }).then(() => {
-  server.listen(3001, () => {
-    console.log('%s listening at 3001'); // eslint-disable-line no-console
-  });
-});
